pyproject.toml
[tool.poetry]
name = "genai-incident-mgmt"
version = "0.1.0"
description = "GenAI incident management system"
authors = ["Your Name <email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.7"
boto3 = "^1.34.0"
crewai = "^0.201.1"
crewai-tools = "^0.75.0"
fastapi = "^0.112.0"
uvicorn = "^0.29.0"
mangum = "^0.17.0"
sentence-transformers = "^3.0.0"
typer = ">=0.9.0,<0.10.0"
aws-cdk-lib = "^2.129.0"
constructs = "^10.3.0"
setuptools = "<81"
pydantic = ">=2.4.2,<3.0.0"
opensearch-py = "^2.4.0"
litellm = "^1.41.0"
python-dotenv = "^1.1.1"
requests-aws4auth = "^1.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
moto = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"